split out uirendering class
