+++ src/ui/ui.cpp
@@ -3,14 +3,10 @@
 #include "bahnhof/ui/tables.h"
 #include "bahnhof/ui/panels.h"
 #include "bahnhof/graphics/rendering.h"
-#include "bahnhof/graphics/graphics.h"
-#include "bahnhof/common/gamestate.h"
 
 
-InterfaceManager::InterfaceManager(Game* newgame)
+InterfaceManager::InterfaceManager(Game* newgame) : game(newgame), uirendering(*this)
 {
-    game = newgame;
-    setuiscale(game->getrendering().getlogicalscale());
     new UI::MainPanel(this, {0,0,180,280});
 }
 
@@ -27,67 +23,14 @@ void InterfaceManager::update(int ms)
 
 void InterfaceManager::render(Rendering* r)
 {
-    int viewheight = screentoui(r->getviewsize()).y;
-    
-    renderscaleruler(r, 20, viewheight-20, 200);
+    int viewheight = uirendering.screentoui(r->getviewsize()).y;
+    uirendering.renderscaleruler(r, 20, viewheight-20, 200);
 
     for(auto pit = panels.rbegin(); pit!=panels.rend(); ++pit)
         (*pit)->render(r);
 
     if(dropdown)
         dropdown->render(r);
-}
-
-void InterfaceManager::renderscaleruler(Rendering* r, int leftx, int lefty, int scalelinelength)
-{
-    float uiscale = getuiscale();
-
-    int textheight = 14;
-    SDL_Color c = getcolorfromstyle(UI::MapOverlay);
-    SDL_SetRenderDrawColor(renderer, c.r, c.g, c.b, c.a);
-    int markersize = 2;
-	r->renderline(uitoscreen(Vec(leftx,lefty)), uitoscreen(Vec(leftx+scalelinelength,lefty)), false);
-	r->renderline(uitoscreen(Vec(leftx,lefty-markersize)), uitoscreen(Vec(leftx,lefty+markersize)), false);
-	r->renderline(uitoscreen(Vec(leftx+scalelinelength,lefty-markersize)), uitoscreen(Vec(leftx+scalelinelength,lefty+markersize)), false);
-    std::string scaletext = std::to_string(int(round(scalelinelength*0.001*150/r->getcamscale()*uiscale))) + " m";
-	rendertext(r, scaletext, {leftx, lefty-textheight, scalelinelength, textheight}, UI::MapOverlay, true);
-}
-
-SDL_Color InterfaceManager::getcolorfromstyle(UI::TextStyle style)
-{
-    SDL_Color color;
-    switch (style)
-    {
-    case UI::Info:
-        color = {0,0,0,255};
-        break;
-    case UI::Highlighted:
-        color = {255,255,255,255};
-        break;
-    case UI::InvertedInfo:
-        color = {255,255,255,255};
-        break;
-    case UI::MapOverlay:
-        color = {255,255,255,255};
-        break;
-    default:
-        std::cout<<"Warning: UI style "<<style<<" not supported in InterfaceManager::rendertext"<<std::endl;
-        break;
-    }
-    return color;
-}
-
-SDL_Rect InterfaceManager::rendertext(Rendering* r, std::string text, SDL_Rect rect, UI::TextStyle style, bool centered)
-{
-    SDL_Color color = getcolorfromstyle(style);
-    float scale = getuiscale();
-    SDL_Rect absrect = uitoscreen(rect);
-    SDL_Rect textrect;
-    if(centered)
-        textrect = r->rendercenteredtext(text, absrect.x+absrect.w*0.5, absrect.y+absrect.h*0.5, color, false, false, absrect.w);
-    else
-        textrect = r->rendertext(text, absrect.x, absrect.y, color, false, false, absrect.w);
-    return screentoui(textrect);
 }
 
 UI::Host* InterfaceManager::getpanelat(Vec screenpos)
@@ -134,7 +77,7 @@ bool InterfaceManager::click(Vec mousepos, int type)
             clickedui = true;
             movingwindow = clickedpanel;
             SDL_Rect movingwindowrect = movingwindow->getglobalrect();
-            movingwindowoffset = screentoui(mousepos)-Vec(movingwindowrect.x, movingwindowrect.y);
+            movingwindowoffset = uirendering.screentoui(mousepos)-Vec(movingwindowrect.x, movingwindowrect.y);
             clickedpanel->click(mousepos, type);
         }
     }
@@ -153,7 +96,7 @@ bool InterfaceManager::leftpressed(Vec mousepos, int mslogic)
         pressedpanel->mousepress(mousepos, mslogic, SDL_BUTTON_LEFT);
     }
     if(movingwindow){
-        movingwindow->move(screentoui(mousepos) - movingwindowoffset);
+        movingwindow->move(uirendering.screentoui(mousepos) - movingwindowoffset);
         return true;
     }
     return false;
@@ -201,54 +144,4 @@ void InterfaceManager::setdropdown(UI::Dropdown* dr)
 Game& InterfaceManager::getgame()
 {
     return *game;
-}
-
-float InterfaceManager::getuiscale()
-{
-    return uiscale;
-}
-
-void InterfaceManager::increaseuiscale()
-{
-    setuiscale(uiscale + 0.2);
-}
-
-void InterfaceManager::decreaseuiscale()
-{
-    setuiscale(uiscale - 0.2);
-}
-
-void InterfaceManager::setuiscale(float newscale)
-{
-    uiscale = std::fmax(0.8, newscale);;
-    int newfontsize = 12*uiscale;
-    setfontsize(newfontsize);
-}
-
-SDL_Rect InterfaceManager::uitoscreen(SDL_Rect uirect)
-{
-    float scale = getuiscale();
-    SDL_Rect screenrect = {int(round(uirect.x*scale)), int(round(uirect.y*scale)), int(round(uirect.w*scale)), int(round(uirect.h*scale))};
-    return screenrect;
-}
-
-SDL_Rect InterfaceManager::screentoui(SDL_Rect screenrect)
-{
-    float scale = getuiscale();
-    SDL_Rect uirect = {int(round(screenrect.x/scale)), int(round(screenrect.y/scale)), int(round(screenrect.w/scale)), int(round(screenrect.h/scale))};
-    return uirect;
-}
-
-Vec InterfaceManager::uitoscreen(Vec pos)
-{
-    float scale = getuiscale();
-    Vec uipos = Vec(pos.x*scale, pos.y*scale);
-    return uipos;
-}
-
-Vec InterfaceManager::screentoui(Vec pos)
-{
-    float scale = getuiscale();
-    Vec uipos = Vec(pos.x/scale, pos.y/scale);
-    return uipos;
 }
